modify user modal

+++ zds/gallery/mixins.py
@@ -48,6 +48,10 @@ class GalleryCreateMixin(GalleryMixin):
         self.users_and_permissions = {user.pk: {'read': True, 'write': True}}
 
 
+class NoMoreUserWithWriteIfLeave(Exception):
+    pass
+
+
 class GalleryUpdateOrDeleteMixin(GalleryMixin):
     def perform_update(self, data):
         """Update gallery information"""
@@ -80,9 +84,21 @@ class GalleryUpdateOrDeleteMixin(GalleryMixin):
         self.gallery.delete()
 
     def perform_leave(self, user):
-        """Remove user. If that user was the last, delete gallery
-        Return True if the gallery was deleted, False otherwise
+        """Remove user.
+        Return True if the gallery was deleted, False otherwise.
+        Fail if the user was the last with write permissions on the gallery.
         """
+        still_one_user_with_write = False
+        for user_pk, user_perms in self.users_and_permissions.items():
+            if user_pk == user.pk:
+                continue
+            if user_perms['write']:
+                still_one_user_with_write = True
+                break
+
+        if not still_one_user_with_write:
+            raise NoMoreUserWithWriteIfLeave()
+
         if user.pk in self.users_and_permissions:
             user_gallery = UserGallery.objects.filter(user=user, gallery=self.gallery).get()
             user_gallery.delete()
